PixTestDaq-improvement for module test
changed again histogram fill to avoid problems with module test.
As previously requested, the same set of histograms(v0) will be filled for each run (doTest).
Warning: this is done even if the input parameters are changed between one run and the other. Pay attention to collect coherent data. To fill another histogram exit pXar and restart.
The Clear button still gives a crash! (as for xray test)