Wrapped RS232 in an object to re-introduce multiple port functionality.
Renamed VoltageMin and VoltageMax to VoltageStart and VoltageStop for IVCurve
Testing. This is to reflect the new ability to do both forward and reverse sweeps.

Version is now working with a Keithley 2410 Power supply on an Ubuntu laptop.

Iseg implementation still needs to be rewritten but I have no hardware for testing it.