Fixed problem for single ROC in settimings pretest
After adding a fix to the set timing pre-test for TBM09, #345 broke it for a single ROC. This fixes that fix. I still think there maybe be a problem with the timing scan for TBM09 if the default timings don't work. But since I don't have a TBM09, I'm not going to worry about it.